+++ src/routes/_/Thing.svelte
@@ -1,9 +1,9 @@
 <script>
-	import { knobby } from '$lib';
+	import * as knobby from '$lib';
 
 	// controls are only rendered when they are being
 	// used; these will disappear if <Thing/> is unmounted
-	const controls = knobby({
+	const controls = knobby.panel({
 		$label: '<Thing> options',
 		a: 1,
 		b: 2